refactor(volumes): split volume view model responsibilities
