+++ ArcBox/ViewModels/Volumes/VolumeTypes.swift
@@ -1,0 +1,14 @@
+/// Detail tab for volumes
+enum VolumeDetailTab: String, CaseIterable, Identifiable {
+    case info = "Info"
+    case files = "Files"
+
+    var id: String { rawValue }
+}
+
+/// Sort field for volumes
+enum VolumeSortField: String, CaseIterable {
+    case name = "Name"
+    case dateCreated = "Date Created"
+    case size = "Size"
+}
